refactor: split Thunks and migrate api file in store
